feat(rpc): remove `v02` and `v03`
