+++ crates/rpc/src/executor.rs
@@ -4,12 +4,12 @@ use pathfinder_common::{ChainId, StarknetVersion};
 use pathfinder_executor::{ClassInfo, IntoStarkFelt};
 use starknet_api::core::PatriciaKey;
 
-use crate::v02::types::request::{
+use crate::types::request::{
     BroadcastedDeployAccountTransaction,
     BroadcastedInvokeTransaction,
     BroadcastedTransaction,
 };
-use crate::v02::types::SierraContractClass;
+use crate::types::SierraContractClass;
 
 pub enum ExecutionStateError {
     BlockNotFound,
@@ -29,7 +29,7 @@ pub(crate) fn map_broadcasted_transaction(
     transaction: &BroadcastedTransaction,
     chain_id: ChainId,
 ) -> anyhow::Result<pathfinder_executor::Transaction> {
-    use crate::v02::types::request::BroadcastedDeclareTransaction;
+    use crate::types::request::BroadcastedDeclareTransaction;
 
     let class_info = match &transaction {
         BroadcastedTransaction::Declare(BroadcastedDeclareTransaction::V0(tx)) => {
