+++ crates/rpc/src/method/add_declare_transaction.rs
@@ -8,7 +8,7 @@ use starknet_gateway_types::request::add_transaction::{
 };
 
 use crate::context::RpcContext;
-use crate::v02::types::request::BroadcastedDeclareTransaction;
+use crate::types::request::BroadcastedDeclareTransaction;
 
 #[derive(Debug)]
 pub enum AddDeclareTransactionError {
@@ -319,13 +319,13 @@ mod tests {
     };
 
     use super::*;
-    use crate::v02::types::request::{
+    use crate::types::request::{
         BroadcastedDeclareTransaction,
         BroadcastedDeclareTransactionV1,
         BroadcastedDeclareTransactionV2,
         BroadcastedDeclareTransactionV3,
     };
-    use crate::v02::types::{
+    use crate::types::{
         CairoContractClass,
         ContractClass,
         DataAvailabilityMode,
@@ -383,7 +383,7 @@ mod tests {
             use super::super::*;
             use crate::dto::serialize::SerializeForVersion;
             use crate::dto::{serialize, DeserializeForVersion};
-            use crate::v02::types::request::BroadcastedDeclareTransactionV1;
+            use crate::types::request::BroadcastedDeclareTransactionV1;
             use crate::RpcVersion;
 
             fn test_declare_txn() -> Transaction {
@@ -480,7 +480,7 @@ mod tests {
 
             use super::super::*;
             use crate::dto::DeserializeForVersion;
-            use crate::v02::types::request::BroadcastedDeclareTransactionV2;
+            use crate::types::request::BroadcastedDeclareTransactionV2;
             use crate::RpcVersion;
 
             fn test_declare_txn() -> Transaction {
