+++ crates/rpc/src/method/get_class_at.rs
@@ -4,7 +4,7 @@ use pathfinder_common::{BlockId, ContractAddress};
 use crate::context::RpcContext;
 use crate::dto;
 use crate::dto::serialize::SerializeForVersion;
-use crate::v02::types::{CairoContractClass, ContractClass, SierraContractClass};
+use crate::types::{CairoContractClass, ContractClass, SierraContractClass};
 
 crate::error::generate_rpc_error_subset!(Error: BlockNotFound, ContractNotFound);
