+++ crates/rpc/src/method/subscribe_events.rs
@@ -255,7 +255,7 @@ mod tests {
     use crate::context::{RpcConfig, RpcContext};
     use crate::jsonrpc::{handle_json_rpc_socket, RpcRouter, CATCH_UP_BATCH_SIZE};
     use crate::pending::PendingWatcher;
-    use crate::v02::types::syncing::Syncing;
+    use crate::types::syncing::Syncing;
     use crate::{v08, Notifications, Reorg, SyncState};
 
     #[tokio::test]
