+++ crates/rpc/src/method/subscribe_new_heads.rs
@@ -179,7 +179,7 @@ mod tests {
     use crate::context::{RpcConfig, RpcContext};
     use crate::jsonrpc::{handle_json_rpc_socket, RpcResponse, RpcRouter, CATCH_UP_BATCH_SIZE};
     use crate::pending::PendingWatcher;
-    use crate::v02::types::syncing::Syncing;
+    use crate::types::syncing::Syncing;
     use crate::{v08, Notifications, Reorg, SubscriptionId, SyncState};
 
     #[tokio::test]
