+++ crates/rpc/src/method/subscribe_pending_transactions.rs
@@ -158,7 +158,7 @@ mod tests {
     use crate::context::{RpcConfig, RpcContext};
     use crate::jsonrpc::{handle_json_rpc_socket, RpcResponse};
     use crate::pending::PendingWatcher;
-    use crate::v02::types::syncing::Syncing;
+    use crate::types::syncing::Syncing;
     use crate::{v08, Notifications, PendingData, SyncState};
 
     #[tokio::test]
