--- crates/rpc/src/v02/method.rs
@@ -1,18 +1,0 @@
-mod block_hash_and_number;
-mod chain_id;
-mod get_block_transaction_count;
-mod get_class;
-mod get_class_at;
-mod get_class_hash_at;
-mod get_nonce;
-mod get_storage_at;
-pub(crate) mod get_transaction_by_block_id_and_index;
-pub(crate) mod get_transaction_by_hash;
-
-pub(crate) use chain_id::chain_id;
-pub(crate) use get_block_transaction_count::get_block_transaction_count;
-pub(crate) use get_class::get_class;
-pub(crate) use get_class_at::get_class_at;
-pub(crate) use get_class_hash_at::get_class_hash_at;
-pub(crate) use get_nonce::get_nonce;
-pub(crate) use get_storage_at::get_storage_at;
