--- crates/rpc/src/v03.rs
@@ -1,1 +1,0 @@
-pub mod method;
