--- crates/rpc/src/v03/method.rs
@@ -1,5 +1,0 @@
-mod get_events;
-pub(crate) mod get_state_update;
-
-pub(crate) use get_events::get_events;
-pub(crate) use get_state_update::get_state_update;
