+++ crates/rpc/src/v06/method.rs
@@ -7,7 +7,6 @@ pub(crate) mod estimate_message_fee;
 mod get_block_with_tx_hashes;
 mod get_block_with_txs;
 mod get_transaction_by_block_id_and_index;
-mod get_transaction_by_hash;
 pub(crate) mod get_transaction_receipt;
 mod get_transaction_status;
 pub(crate) mod simulate_transactions;
@@ -24,7 +23,6 @@ pub(crate) use estimate_message_fee::estimate_message_fee;
 pub(crate) use get_block_with_tx_hashes::get_block_with_tx_hashes;
 pub(crate) use get_block_with_txs::get_block_with_txs;
 pub(crate) use get_transaction_by_block_id_and_index::get_transaction_by_block_id_and_index;
-pub(crate) use get_transaction_by_hash::get_transaction_by_hash;
 pub(super) use get_transaction_receipt::get_transaction_receipt;
 pub(crate) use get_transaction_status::get_transaction_status;
 pub(crate) use simulate_transactions::simulate_transactions;
