+++ crates/rpc/src/v06/method/add_invoke_transaction.rs
@@ -4,7 +4,7 @@ use starknet_gateway_types::error::SequencerError;
 
 use crate::context::RpcContext;
 use crate::felt::RpcFelt;
-use crate::v02::types::request::BroadcastedInvokeTransaction;
+use crate::types::request::BroadcastedInvokeTransaction;
 
 #[derive(serde::Deserialize, Debug, PartialEq, Eq)]
 #[serde(tag = "type")]
@@ -191,8 +191,8 @@ mod tests {
     use pathfinder_common::{ResourceAmount, ResourcePricePerUnit, Tip, TransactionVersion};
 
     use super::*;
-    use crate::v02::types::request::BroadcastedInvokeTransactionV1;
-    use crate::v02::types::{DataAvailabilityMode, ResourceBound, ResourceBounds};
+    use crate::types::request::BroadcastedInvokeTransactionV1;
+    use crate::types::{DataAvailabilityMode, ResourceBound, ResourceBounds};
 
     fn test_invoke_txn() -> Transaction {
         Transaction::Invoke(BroadcastedInvokeTransaction::V1(
@@ -331,7 +331,7 @@ mod tests {
     #[tokio::test]
     #[ignore = "gateway 429"]
     async fn duplicate_transaction() {
-        use crate::v02::types::request::BroadcastedInvokeTransactionV1;
+        use crate::types::request::BroadcastedInvokeTransactionV1;
 
         let context = RpcContext::for_tests();
         let input = BroadcastedInvokeTransactionV1 {
@@ -372,7 +372,7 @@ mod tests {
     #[ignore = "gateway 429"]
     // https://external.integration.starknet.io/feeder_gateway/get_transaction?transactionHash=0x41906f1c314cca5f43170ea75d3b1904196a10101190d2b12a41cc61cfd17c
     async fn duplicate_v3_transaction() {
-        use crate::v02::types::request::BroadcastedInvokeTransactionV3;
+        use crate::types::request::BroadcastedInvokeTransactionV3;
 
         let context = RpcContext::for_tests_on(pathfinder_common::Chain::SepoliaIntegration);
         let input = BroadcastedInvokeTransactionV3 {
